Organize pages into chapters folder
