+++ src/applications/income-and-asset-statement/config/chapters/01-applicant-information/applicantIsVeteran.js
