+++ src/applications/income-and-asset-statement/config/chapters/01-applicant-information/index.js
@@ -1,0 +1,10 @@
+import applicantInformation from './applicantInformation';
+import applicantIsVeteran from './applicantIsVeteran';
+
+export default {
+  title: 'Applicant information',
+  pages: {
+    applicantInformation,
+    applicantIsVeteran,
+  },
+};
