+++ src/applications/income-and-asset-statement/config/chapters/02-claimant-information/claimantType.js
