+++ src/applications/income-and-asset-statement/config/chapters/02-claimant-information/index.js
@@ -1,0 +1,10 @@
+import claimantInformation from './claimantInformation';
+import claimantType from './claimantType';
+
+export default {
+  title: 'Claimant information',
+  pages: {
+    claimantInformation,
+    claimantType,
+  },
+};
